[10.x] Move Symfony events dispatcher registration to `Console\Kernel`
