+++ src/Illuminate/Foundation/Console/Kernel.php
@@ -7,6 +7,8 @@ use Closure;
 use DateTimeInterface;
 use Illuminate\Console\Application as Artisan;
 use Illuminate\Console\Command;
+use Illuminate\Console\Events\CommandFinished;
+use Illuminate\Console\Events\CommandStarting;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Contracts\Console\Kernel as KernelContract;
 use Illuminate\Contracts\Debug\ExceptionHandler;
@@ -18,6 +20,10 @@ use Illuminate\Support\Env;
 use Illuminate\Support\InteractsWithTime;
 use Illuminate\Support\Str;
 use ReflectionClass;
+use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Console\Event\ConsoleCommandEvent;
+use Symfony\Component\Console\Event\ConsoleTerminateEvent;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Finder\Finder;
 use Throwable;
 
@@ -38,6 +44,13 @@ class Kernel implements KernelContract
      * @var \Illuminate\Contracts\Events\Dispatcher
      */
     protected $events;
+
+    /**
+     * The symfony event dispatcher instance.
+     *
+     * @var \Symfony\Contracts\EventDispatcher\EventDispatcherInterface|null
+     */
+    protected $symfonyDispatcher = null;
 
     /**
      * The Artisan application instance.
@@ -104,6 +117,8 @@ class Kernel implements KernelContract
 
         $this->app = $app;
         $this->events = $events;
+
+        $this->rerouteSymfonyCommandEvents();
 
         $this->app->booted(function () {
             $this->defineConsoleSchedule();
@@ -417,6 +432,10 @@ class Kernel implements KernelContract
             $this->artisan = (new Artisan($this->app, $this->events, $this->app->version()))
                                     ->resolveCommands($this->commands)
                                     ->setContainerCommandLoader();
+
+            if (! $this->app->runningUnitTests()) {
+                $this->artisan->setDispatcher($this->symfonyDispatcher);
+            }
         }
 
         return $this->artisan;
@@ -464,5 +483,38 @@ class Kernel implements KernelContract
     protected function renderException($output, Throwable $e)
     {
         $this->app[ExceptionHandler::class]->renderForConsole($output, $e);
+    }
+
+    /**
+     * Re-route the Symfony command events to their Laravel counterparts.
+     *
+     * @internal
+     *
+     * @return void
+     */
+    protected function rerouteSymfonyCommandEvents()
+    {
+        $this->symfonyDispatcher = new EventDispatcher;
+
+        $this->symfonyDispatcher->addListener(ConsoleEvents::COMMAND, function (ConsoleCommandEvent $event) {
+            $this->events->dispatch(
+                new CommandStarting(
+                    $event->getCommand()->getName(),
+                    $event->getInput(),
+                    $event->getOutput(),
+                )
+            );
+        });
+
+        $this->symfonyDispatcher->addListener(ConsoleEvents::TERMINATE, function (ConsoleTerminateEvent $event) {
+            $this->events->dispatch(
+                new CommandFinished(
+                    $event->getCommand()->getName(),
+                    $event->getInput(),
+                    $event->getOutput(),
+                    $event->getExitCode(),
+                )
+            );
+        });
     }
 }
